Add custom libinit
This lays the framework to set variant specific values in the device
trees. This is based on the work done for msm8916 samsung. But has
been extended to support LG and how certain things are formatted.
It also supports GSM, CDMA, WIFI common props but they are currently
unused but can be expanded if other devices are added in the future.